requirements.txt
python-telegram-bot==13.15
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
aiohttp==3.8.5
requests==2.31.0